+++ frontend/src/pages/RepositoryListing/components/ResultList/Icons/StarIcon.tsx
