+++ frontend/src/pages/RepositoryListing/components/ResultList/Icons/WatcherIcon.tsx
